refactor(session): split bus query responsibilities
